==> 1-Data_Generation.ipynb <==
# Welcome to the Data Generation Notebook

### Run this Notebook to Store CSV's - Which will be later used by the Models
Statistical Analysis and Machine Learning Models used for Predictions (see Notebooks)
1. Hot Numbers Frequency Analysis (most drawn)
2. Cold Numbers Frequency Analysis (least drawn)
3. Numbers Longest not drawn (bases on date of drawing they last appeared)
4. Monte Carlo Simulation (just random fun for me, 2 Lotto Simulations run against each other till they match)
5. LSTM-Model - Long-Short-Term-Memory Model (recurrent neural network (RNN))
6. Random Forest Model - Machine Learning Model

# Data Collection and CSV Creation + Cleaning

# Import Libraries needed
import pandas as pd
import matplotlib.pyplot as plt


### Get the Data from Johannesfriedrich on Github! Thanks for Providing them in JSON Format and always refresh the Set 

- There are 2 Version of the Data you can get i will load data_json and do my own Data Cleaning and EDA
- The first Version seems not to have the "Superzahl" thats why i clean and Shape the Data by myself

Special Thanks for Scraping and providing always the newest Numbers for us. Check out his Github! https://github.com/JohannesFriedrich

# get the Lotto Numbers from the JSON Files in the github repository
data_json_1 = "https://johannesfriedrich.github.io/LottoNumberArchive/Lottonumbers_tidy_complete.json"
data_json = "https://johannesfriedrich.github.io/LottoNumberArchive/Lottonumbers_complete.json"


# read in the json file from the provided URL
df = pd.read_json(data_json)

### Normalize the data because only 1 column "data" which holds dictionaries

# normalize the json file into columns

df = pd.json_normalize(df['data'])

### Superzahl was Introduced in the 2000's so i will make a 2nd Dataframe only with drawings which contained the Superzahl

df_superzahl = df.dropna(subset=["Superzahl"])


### Write the Data into a CSV called frequency_data and Store 1 CSV named superzahl


# Write the DataFrames to CSV files
df.to_csv('data/frequency_data.csv', index=False)
df_superzahl.to_csv('data/superzahl_data.csv', index=False)

# Continue with the Next Notebook for the first Predictions!

If you are only interested in 1 Method to Forecast Numbers you can directly skip to this one
