==> 4-Long_Time_no_see.ipynb <==
## 3. Prediction - Longest time not Drawn - Based on Date Column

### About this Method and the assumption we have behind it:
- We assume the 6 Numbers which we did not see the longest time should be drawn in short future
- We try our Luck and pick the 6 Numbers which we did not see the longest time

### The Statistical Assumption behind this
- In a perfect World with infinite Lottery Games all Numbers should always be drawn equally often at one point in time
- We assume that it's on the time, that some of the longest not drawn numbers appear on the next drawings

For this 3 Prediction we Take the 6 longest not Drawn Numbers


# Import Libraries needed
import pandas as pd
import matplotlib.pyplot as plt
import ast
import numpy as np
import pickle

plt.style.use('https://github.com/dhaitz/matplotlib-stylesheets/raw/master/pitayasmoothie-dark.mplstyle')

### Load the Data we Stored in the data Folder with Notebook 1


# Load the CSV file into a DataFrame
frequency_data = pd.read_csv("data/frequency_data.csv")


### Now we count the Frequency of the Numbers which were drawn since the beginning of the Records

frequency_data.head()

### Convert date from String to Datetime for Evaluation
And further Data Processing for our needs

# convert the string to datetime
frequency_data["date"] = pd.to_datetime(frequency_data["date"], format='%d.%m.%Y')

# Convert the Lottozahl column to a list of integers
frequency_data["Lottozahl"] = frequency_data["Lottozahl"].apply(ast.literal_eval)

# Explode the list into individual rows
frequency_data = frequency_data.explode("Lottozahl")

# Convert the Lottozahlen to integers
frequency_data["Lottozahl"] = frequency_data["Lottozahl"].astype(int)

### Here we Group by Lottozahl and Date

# Group data to find the last drawn date for each number

last_drawn_date = frequency_data.groupby("Lottozahl")["date"].max()

# Sort this list to get the least_recently_drawn Numbers
least_recently_drawn = last_drawn_date.sort_values().head(6)


# Predictions for the 3 Lottery Ticket Field! - Longest Time not seen

### DISCLAIMER:
- This is a Fun Project to Showcase some IT Skills
- Lottery is a total statistically Random Game
- Do not use or let you inspire by this Prediction!
- Gambling can make you addicted! Only Play with Caution!
- I claim no rights, that my outputs would even be near the drawn value

### About the Predictions:
- In Field 3 we go for the longest Time no See Numbers each Lottery Ticket we play
- The Superzahl only appears once per Lottery Ticket - we will take Random 7, 5 or 4 for each Lottery Ticket
- This is a common Lottery Strategy - even if you win with this Numbers, you would have to share the Prize Pool with many other People
- We will do more complex Prediction Strategies in later Notebooks
- This Prediction will vary more from Week to Week than the first 2, because when a Number gets drawn it refreshes its date and will not show up here for a longer time

# Format the first 6 Numbers for better Output
formatted_numbers = ", ".join(map(str, least_recently_drawn.index))



# Print your Lucky Numbers
print("Your Lucky Numbers For Field 3 (Longest Time not Drawn):",formatted_numbers," Superzahl:", np.random.choice([4,7]))

## Optional - Store the Numbers for BONUS Notebook 8 in the End!!!

# Store the List in a Pickle file

with open("stored_predictions/time_numbers.pickle", "wb") as f:
    pickle.dump(formatted_numbers, f)
