==> 5-Monte_Carlo_simulation.ipynb <==
# 4. Prediction - Monte Carlo Simulation (Cross Validation :-)  ) 

## Here we let 2 Monte Carlo Simulations draw 6 Numbers each and let them do this so long till both Draws Match all 6 Numbers

This makes no Sense at all and Adds more Random to the Number which were drawn. But i think it's a funny way to let it generate my total Random Prediction Field
which underlies no Sense at all

### About the Monte Carlo Simulation
This code is Simulating the drawing of the 6 aus 49 Lottozahlen without the Superzahl!!!

We just let 2 Simulations run against each other and Check if all 6 Numbers are the same after each round of drawing

If both Numbers math on all 6 Positions, we take them as our Random guess for 1 Lottery field 

Hint: Bonus Notebook 8 uses this Concept as a fun way to validate our other 6 predictions ;)

And! We could never know, how long the Code will run till 2 Matching Predictions were found. So be patient when predicting with this Method ( in my test runs it never took longer than 6 Minutes and mostly under 10.000.000 Simulations. Not even near 140 Mio. trys like they say in the commercial Chance 1:140.000.000) 

If you observe other behaviour here feel free to let me know :)

###

# import libraries
import random
import numpy as np
import pickle

### Define some Functions to do the Prediction



# draw numbers function draws us random 6 Numbers in range from 0-49

def draw_numbers():
    return random.sample(range(1,50),6)


# Function Customized for our Needs of the Monte Carlo Function

def monte_carlo_simulation():
    anzahl_simulationen = 0
    while True: # only exits when the return statement is reached
        anzahl_simulationen += 1 # increase for each iteration
        ziehung1 = draw_numbers()
        ziehung2 = draw_numbers()

        if set(ziehung1) == set(ziehung2): # Set orders Numbers and if all 6 are equal we print the predictions
            print(f"Es waren {anzahl_simulationen:,} Simulationen notwendig um die Predictions {ziehung1} zu generieren. Viel Erfolg!") # Print out how often we Simulated Numbers
            print("Your Lucky Numbers For Field 4 (Monte Carlo Random):",ziehung1," Superzahl:", np.random.choice([4,7,5]))
            return ziehung1

### Make our Monte Carlo Simulation Predictions for Field 4 (Monte Carlo Random)

ziehung1 = monte_carlo_simulation()

### Well this was a Short Notebook! But the Code runs the longest of all. We achieved our Prediction for the 4 Field we want to play

Monte Carlo Simulation is a statistical Method where you Simulate Random Examples of Combinations or Numbers and you can Validate how much hit your Numbers would have had if the game was Simulated a certain amount of Time.

If you want to know more about this, feel free to visit: https://www.ibm.com/topics/monte-carlo-simulation#:~:text=One%20simple%20example%20of%20a,probability%20of%20a%20particular%20outcome.

## Optional - Store the Numbers for BONUS Notebook 8 in the End!!!

# Store the List in a Pickle file

with open("stored_predictions/monte_carlo.pickle", "wb") as f:
    pickle.dump(ziehung1, f)
