==> 7-Random_Forest.ipynb <==
## 6. Prediction - Random Forest Model!

### About this Method and the assumption we have behind it:
- Random Forest is kind of a Binary Decision Tree which picks 1 Feature at each Split and tries to find the best splits on each Node
- We try our Luck and let this Kind of Model predict our Numbers for Prediction Nr. 6
- This is a Machine Learning Model which works on linear Regression as well as Classification Problems

This is our Final Prediction then we have our 6 Fields we play each week. 


# Import Libraries needed
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error
import numpy as np
import matplotlib.pyplot as plt
from sklearn.model_selection import GridSearchCV
import pickle
plt.style.use('https://github.com/dhaitz/matplotlib-stylesheets/raw/master/pitayasmoothie-dark.mplstyle')

### Load the Data we Stored in the data Folder with Notebook 1


# Load the CSV file into a DataFrame
frequency_data = pd.read_csv("data/frequency_data.csv")


### Data Preprocessing so we can use it as Input for the Model

# convert the string to datetime
frequency_data["date"] = pd.to_datetime(frequency_data["date"], format='%d.%m.%Y')

# Convert the Lottozahl column to a list of integers
frequency_data["Lottozahl"] = frequency_data["Lottozahl"].apply(lambda x: eval(x))

# Expand the Lottozahl column into seperate columns
df_expanded = pd.DataFrame(frequency_data["Lottozahl"].tolist(), columns=['Lottozahl_1', 'Lottozahl_2', 'Lottozahl_3', 'Lottozahl_4', 'Lottozahl_5', 'Lottozahl_6'])

# Create the final Dataframe
df_final = pd.concat([frequency_data.drop(columns= "Lottozahl"), df_expanded], axis=1)


# Cast the Date into own columns
df_final["day"] = df_final["date"].dt.day
df_final["month"] = df_final["date"].dt.month
df_final["year"] = df_final["date"].dt.year

# Drop the Superzahl Column (we only Predict the 6 Lotto Numbers)
df_final.drop(["Superzahl", "id","date"], axis=1, inplace=True)

df_final.head()

### Define X and y - X are our Input Features y is our Target Variable
after that we do a Train test split and use a param_grid to Cross Validate our Results

# Assuming df is your DataFrame containing the data
X = df_final[['day', 'month', 'year']]  # Features: day, month, year
y = df_final[['Lottozahl_1', 'Lottozahl_2', 'Lottozahl_3', 'Lottozahl_4', 'Lottozahl_5', 'Lottozahl_6']]  # Target variables: Lotto numbers

# Train Test Split
X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.2, random_state=42)

### Define the Param Grid to find the best Hyperparameters

# Define the parameter grid
param_grid = {
    'n_estimators': [50, 100, 150],  # Number of trees in the forest
    'max_depth': [None, 10, 20],  # Maximum depth of the tree
    'min_samples_split': [2, 5, 10],  # Minimum number of samples required to split an internal node
    'min_samples_leaf': [1, 2, 4]  # Minimum number of samples required to be at a leaf node
}


### Train the Random Forest Model with a Grid Search

# define the model
rf = RandomForestRegressor(random_state = 42)

# Perform grid search cross-validation
grid_search = GridSearchCV(estimator=rf, param_grid=param_grid, cv=5, scoring="neg_mean_squared_error", n_jobs=-1, verbose=2)
grid_search.fit(X_train, y_train)

### Use the Best Params found in Grid Search for the final Random Forest Model

# Best Param out of Grid Search Cross Validation
best_params = grid_search.best_params_
best_score = grid_search.best_score_


print("Best Parameters:", best_params)
print("Best Negative Mean Squared Error: ", best_score)

# Get the best model
best_rf_model = grid_search.best_estimator_

# Evaluate the best model
y_pred_best = best_rf_model.predict(X_test)
mse_best = mean_squared_error(y_test,y_pred_best)
print("Mean Squared Error for Best Model:", mse_best)



## Adjust the Date here for the Date for which u want to have Numbers!

Adjust the date in the new_dates Variable to the Date where you want to have the Predicted Numbers for the drawing


# Change the Day of the Drawing here
new_dates = np.array([[4, 5, 2024]])  # CHANGE THE DATE HERE FOR WHICH DAY U WANT THE PREDICTION
# Create a DataFrame
new_dates = pd.DataFrame(new_dates, columns=['day', 'month', 'year'])

# Predict lotto numbers for new dates using the best model
new_predictions_best = best_rf_model.predict(new_dates)
new_predictions_int = new_predictions_best[0].astype(int).tolist()
print("Predicted Lotto Numbers for New Dates using Best Random Forrest Model:", new_predictions_int)

## Optional - Store the Numbers for BONUS Notebook 8 in the End!!!

# Store the List in a Pickle file

with open("stored_predictions/random_forest.pickle", "wb") as f:
    pickle.dump(new_predictions_int, f)

# Predictions for the 6 Lottery Ticket Field!

### DISCLAIMER:
- This is a Fun Project to Showcase some IT Skills
- Lottery is a total statistically Random Game
- Do not use or let you inspire by this Prediction!
- Gambling can make you addicted! Only Play with Caution!
- I claim no rights, that my outputs would even be near the drawn value
