==> 8-Bonus_Validation_Monte_Carlo.ipynb <==
# Bonus Notebook 8! Validate your Predictions after 1 Million Lotto Drawings
## Monte Carlo Simulation for Validation of my Predicted Numbers
- We will check how often Each Model would have won based on Simulating the Lotto 6 aus 49 Game
- We even count how much "Richtige" or "Treffer" we have for each Model in Percent
- You can only run this, if you have run all previous Notebooks and stored the Pickle Files at the End!

# import libraries
import random
import pandas as pd
import pickle

## Load all the Pickle Files we Stored for our Models

# Load the pickle files
with open("stored_predictions/hot_numbers.pickle", "rb") as f:
    hot_numbers = pickle.load(f)

with open("stored_predictions/cold_numbers.pickle", "rb") as f:
    cold_numbers = pickle.load(f)

with open("stored_predictions/time_numbers.pickle", "rb") as f:
    long_time_no_see = pickle.load(f)

with open("stored_predictions/monte_carlo.pickle", "rb") as f:
    monte_carlo = pickle.load(f)

with open("stored_predictions/lstm_model.pickle", "rb") as f:
    lstm_model = pickle.load(f)

with open("stored_predictions/random_forest.pickle", "rb") as f:
    random_forest = pickle.load(f)

### Some Pickle Files i stored accidently as String or other Data Types, we will cast everything to a List

long_time_no_see

# For hot Numbers
# Split the string by commas and convert the substrings to integers
hot_numbers_list = [int(num) for num in hot_numbers.split(",")]

# For Cold Numbers
# Split the string by commas and convert the substrings to integers
cold_numbers_list = [int(num) for num in cold_numbers.split(",")]

# long time no see is in correct format
long_time_list = [int(num) for num in long_time_no_see.split(",")]

### Custom Function to evaluate our winning Chances with our Predicted Numbers
- How often would we have won with our Predicted Numbers in 1 Million simulations? 
- We Compare our Predicted Number with Random Drawings
- You could Say we make our own Lotto Drawing


# Draw Random Numbers

def draw_numbers():
    return random.sample(range(50),6)

# define the monte carlo simulation with 2 parameters

def monte_carlo_simulation(gewinnzahlen, anzahl_simulationen):
    treffer_statistik = {0:0, 1:0,2:0,3:0,4:0,5:0,6:0}

    for _ in range(anzahl_simulationen):
        gezogene_zahlen= draw_numbers()
        treffer = len(set(gezogene_zahlen) & set(gewinnzahlen))
        treffer_statistik[treffer] += 1

    treffer_prozent = {treffer: (anzahl/ anzahl_simulationen) * 100 for treffer, anzahl in treffer_statistik.items()}

    return treffer_prozent

### Validate our Predictions

### Copy your Predictions from the previous Notebooks into the Lists for Validation

# Main Function

def main():
    anzahl_simulationen = 1000000

    # Liste meiner Vorhersagen und Modelle

    # Copy in your Predictions from the Previous Notebook here to "validate" them somehow
    modelle = {
        "Hot Numbers": hot_numbers_list, # Hot Numbers Predictions
        "Cold Numbers": cold_numbers_list, # Cold Numbers Predictions
        "Long Time no see": long_time_list, # Long time no see Predictions
        "Monte Carlo Random": monte_carlo, # Monte Carlo Random Prediction
        "LSTM-Model": lstm_model, # Long Short Term Memory Model
        "Random-Forest": random_forest, # Random Forest Model

    }

    for modell, vorhersage in modelle.items():
        treffer_prozent = monte_carlo_simulation(vorhersage, anzahl_simulationen)

        print(f"\nTrefferstatistik für {modell} nach {anzahl_simulationen} Simulationen:")
        for treffer, prozent in treffer_prozent.items():
            print(f"{treffer} Richtige: {prozent:.4f}% ")

## Bonus! Validate our Predictions after 1 Million Lottozahlen Drawings! How often would we have won?

if __name__ == "__main__":
    main()

# Thanks for Reviewing my Lottery Prediction Notebook! I hope you had fun and got inspired how a real IT-Nerd is playing Lotto :)
### Leave a Like and hit me up, if you have further Questions
